Create GlobalVariables.sol
This smart contract uses several global functions in the message context, including blockhash, block, tx, and msg. These global functions provide access to information about the current block, transaction, and message. For example, block.number returns the current block number, and msg.sender returns the address of the sender of the current message.

The constructor function of the contract is executed when the contract is deployed, and it uses these global functions to set the values of the public variables in the contract. These variables can then be accessed by external clients to retrieve information about the current transaction and block.